cloudProjectBackend/requirements.txt
asgiref==3.5.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.3
daphne==4.0.0
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==5.0.0
incremental==22.10.0
Markdown==3.4.1
mysqlclient==2.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.1.0
pytz==2022.6
requests==2.28.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
urllib3==1.26.12
zipp==3.10.0
zope.interface==5.5.1

xhtml2pdf~=0.2.8